refactor: refactor the cosmos::grpc module to cosmos::client
