+++ src/cosmos/client/error.rs
@@ -1,6 +1,6 @@
 //! Error for grpc clients
 
-use crate::cosmos::grpc::error::Error::{InvalidMnemonic, MnemonicParseFailed};
+use crate::cosmos::client::error::Error::{InvalidMnemonic, MnemonicParseFailed};
 use bip39::Error as Bip39Error;
 use cosmrs::bip32::Error as Bip32Error;
 use cosmrs::proto::prost::EncodeError;
