--- src/cosmos/grpc/mod.rs
@@ -1,5 +1,0 @@
-//! Holds GRPC library.
-
-pub mod account;
-pub mod client;
-pub mod error;
